Ajout de l'interaction avec elements + fix + classement
