+++ Assets/Resources/Prefabs/Elements/Cristals/IslandCore.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 1b35c574ab6500043b5d45b0a9e0e17f
-timeCreated: 1454638812
+guid: 30fe82cf2377fb843bccc85cc7aeb68d
+timeCreated: 1454735278
 licenseType: Free
 NativeFormatImporter:
   userData:
