+++ Assets/Resources/Prefabs/Player/Player.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4a23a3fe9845e874b9eac923e1375c83
-timeCreated: 1451935736
+guid: d59cf0d48056bda459bebed63029bb2e
+timeCreated: 1454735303
 licenseType: Free
 NativeFormatImporter:
   userData:
