+++ Assets/Resources/Scripts/Class/Chunk.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 8cf30906fa422dc43943387c2f3f9b82
-timeCreated: 1453293308
+guid: 0c22fc9f3268db345a9b03c45612cc33
+timeCreated: 1454735195
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
