+++ Assets/Resources/Scripts/Class/Entity.cs
@@ -42,7 +42,7 @@ public class Entity
     /// <summary>
     /// Instancie l'entite dans le monde. (Must be server!)
     /// </summary>
-    public void Spawn()
+    public virtual void Spawn()
     {
         this.prefab = GameObject.Instantiate(this.prefab, this.prefab.transform.position, this.prefab.transform.rotation) as GameObject;
         NetworkServer.Spawn(this.prefab);
@@ -51,7 +51,7 @@ public class Entity
     /// <summary>
     /// Instancie l'entite dans le monde avec une position. (Must be server!)
     /// </summary>
-    public void Spawn(Vector3 pos)
+    public virtual void Spawn(Vector3 pos)
     {
         this.prefab = GameObject.Instantiate(this.prefab, pos, this.prefab.transform.rotation) as GameObject;
         NetworkServer.Spawn(this.prefab);
@@ -60,7 +60,7 @@ public class Entity
     /// <summary>
     /// Instancie l'entite dans le monde avec une position et un parent. (Must be server!)
     /// </summary>
-    public void Spawn(Vector3 pos, Transform parent)
+    public virtual void Spawn(Vector3 pos, Transform parent)
     {
         this.prefab = GameObject.Instantiate(this.prefab, pos, this.prefab.transform.rotation) as GameObject;
         this.prefab.transform.parent = parent;
@@ -70,7 +70,7 @@ public class Entity
     /// <summary>
     /// Instancie l'entite dans le monde avec une position et une rotation. (Must be server!)
     /// </summary>
-    public void Spawn(Vector3 pos, Quaternion rot)
+    public virtual void Spawn(Vector3 pos, Quaternion rot)
     {
         this.prefab = GameObject.Instantiate(this.prefab, pos, rot) as GameObject;
         NetworkServer.Spawn(this.prefab);
@@ -79,7 +79,7 @@ public class Entity
     /// <summary>
     /// Instancie l'entite dans le monde avec une position et une rotation et un parent. (Must be server!)
     /// </summary>
-    public void Spawn(Vector3 pos, Quaternion rot, Transform parent)
+    public virtual void Spawn(Vector3 pos, Quaternion rot, Transform parent)
     {
         this.prefab = GameObject.Instantiate(this.prefab, pos, rot) as GameObject;
         this.prefab.transform.parent = parent;
