+++ Assets/Resources/Scripts/Class/Entity.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 26bf0e3b788f8c24a8ddf8597e7e9f93
-timeCreated: 1453059710
+guid: b4e350b0132930d48951e0aefc9df464
+timeCreated: 1454735218
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
