+++ Assets/Resources/Scripts/Class/IslandCore.cs
@@ -27,13 +27,36 @@ public class IslandCore : Element
         this.level_portal = 0;
     }
 
-    public IslandCore(Team team, int level_attack, int level_prod, int level_portal) : base()
+    public IslandCore(IslandCore cristal) : base(cristal)
+    {
+        this.t = cristal.T;
+        this.level_upgrade = cristal.level_upgrade;
+        this.level_attack = cristal.level_attack;
+        this.level_prod = cristal.level_prod;
+        this.level_portal = cristal.level_portal;
+    }
+
+    public IslandCore(int id, GameObject prefab, Team team, int level_attack, int level_prod, int level_portal) : base(id, 100, prefab, 1)
     {
         this.t = team;
         this.level_attack = level_attack;
         this.level_portal = level_portal;
         this.level_prod = level_prod;
         this.level_upgrade = level_prod + level_portal + level_attack;
+    }
+
+    // Methods
+    /// <summary>
+    /// Reset le cristal lorsqu'il n'a plus de pv
+    /// </summary>
+    protected override void Kill()
+    {
+        this.t = Team.Neutre;
+        this.level_upgrade = 0;
+        this.level_attack = 0;
+        this.level_prod = 0;
+        this.level_portal = 0;
+        base.life = base.lifeMax;
     }
 
     // Getters & Setters
