+++ Assets/Resources/Scripts/Database/EntityDatabase.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 84bea1c3dabb0fe44ac2ce923923fb6f
-timeCreated: 1453225483
+guid: ef27ea396469115439946e409bb9a60b
+timeCreated: 1454735228
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
