--- Assets/Resources/Scripts/MonoBehaviour.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 822e004bbdf1c604696837e8f0f129c8
-folderAsset: yes
-timeCreated: 1453463513
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
