+++ Assets/Resources/Scripts/Networking/SyncElement.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 006a24135e9072947934fdc1453523c6
-timeCreated: 1454172892
+guid: c33762e7fc8089047a698995df24ad81
+timeCreated: 1454735321
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
