+++ Assets/Resources/Scripts/Player/InputManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 3c968bc8b622f4261917cce080f6705d
-timeCreated: 1452001623
+guid: 51b5a2cb07c6261429365d1ed7eeface
+timeCreated: 1454735241
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
