+++ Assets/Resources/Scripts/Player/Inventory.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b38c4ee3977b63549a9cf5478bbc4870
-timeCreated: 1453463599
+guid: a6a941600e2cb0e4aa248b9c1caeabef
+timeCreated: 1454735246
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
