--- Assets/Resources/Sprites/Icons.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 06274fa5fd1967f42ba5c12b594a8348
-folderAsset: yes
-timeCreated: 1454676400
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
